tests/requirements.txt
beautifulsoup4
cssutils
gprof2dot
pytest
pytest-cov
pytest-error-for-skips
pytest-html
pytest-profiling
pytest-retry
pytest-sugar
python-dotenv
